Finished adding multipliers back in. Fixed a bug that was preventing SemGen from correctly saving processes.

+++ src/semsim/reading/SemSimOWLreader.java
@@ -651,7 +651,7 @@ public class SemSimOWLreader extends ModelReader {
 			if(ax.equalsIgnoreAnnotations(axiom)){
 				if(!ax.getAnnotations(annprop).isEmpty()){
 					OWLLiteral litval = (OWLLiteral) ax.getAnnotations(annprop).toArray(new OWLAnnotation[]{})[0].getValue();
-					val = (int)litval.parseDouble();
+					val = litval.parseInteger();
 				}
 			}
 		}
